Codebase/requirements.txt
streamlit
torch
torchvision
pillow
numpy
scikit-learn
google-generativeai
python-dotenv
requests
pandas
plotly